@@ -7,18 +7,18 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.15.0-116.0.dev <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
 
-  cupertino_icons: ^1.0.0
-  http: 0.12.1
-  connectivity: 2.0.2
-  shared_preferences: 0.5.12+4
-  provider: 4.3.3
+  cupertino_icons: ^1.0.3
+  http: ^0.13.4
+  connectivity: ^3.0.6
+  shared_preferences: ^2.0.8
+  provider: ^6.0.1
 
 
 dev_dependencies:
