@@ -5,18 +5,18 @@ import 'package:flutter_mvvm_boilerplate/utils/shared_preferences_helper.dart';
 /// Always create a getter and setter for variables.
 /// Direct initialisation and access is not recommended.
 class SingletonConstants {
-  static SingletonConstants _instance;
+  static SingletonConstants? _instance;
 
   factory SingletonConstants() => _instance ??= new SingletonConstants._();
 
   SingletonConstants._();
 
-  String _baseUrl;
-  String _authToken;
+  String? _baseUrl;
+  String? _authToken;
 
-  String getBaseUrl() => _baseUrl != null ? _baseUrl : ApiConstants.SERVER_BASE_URL;
+  String? getBaseUrl() => _baseUrl != null ? _baseUrl : ApiConstants.SERVER_BASE_URL;
 
   void setBaseUrl(String baseURL) => _baseUrl = baseURL;
 
-  Future<String> get getToken async => _authToken != null ? _authToken : await SharedPreferencesHelper.getAuthTokenWithNullCheck();
+  Future<String?> get getToken async => _authToken != null ? _authToken : await SharedPreferencesHelper.getAuthTokenWithNullCheck();
 }
