@@ -9,7 +9,7 @@ class ApiConstants {
   ///NOTE: DO NOT TWEAK
   ///Assigning Application's base url which app is going to use. Can use URL from custom page or from SERVER_BASE_URL.
   ///Application have Safe checks so custom URL page will never show if isProdBuild flag in AppConstants is set as true
-  static String appBaseURL =
+  static String? appBaseURL =
       AppConstants.isCustomURLBuild && !AppConstants.isProdBuild ? SingletonConstants().getBaseUrl() : SERVER_BASE_URL;
 
   //Login end point
@@ -20,4 +20,4 @@ class ApiConstants {
 ///NOTE: DO NOT TWEAK
 ///Getter to fetch baseURL easily.
 ///This will be available anywhere on APP level.
-String get baseURL => ApiConstants.appBaseURL;
+String? get baseURL => ApiConstants.appBaseURL;
